feat(#15): extract copied code
