+++ src/routes/entry/[id]/+page.svelte
@@ -101,7 +101,7 @@
 </svelte:head>
 
 <div class="flex h-screen flex-col bg-white">
-`  <main class="flex flex-1 flex-col overflow-y-auto px-4 py-2.5">
+  ` <main class="flex flex-1 flex-col overflow-y-auto px-4 py-2.5">
     <div class="flex w-full flex-col gap-5 pt-2.5 pb-24">
       <!-- Title -->
       <h1 class="text-center text-[22px] leading-7 font-normal text-black">Edit your mood entry</h1>
